Make `nd_plen` to be int to reduce `rb_long2int` usage

+++ rubyparser.h
@@ -805,7 +805,7 @@ typedef struct RNode_ARGS_AUX {
     NODE node;
 
     ID nd_pid;
-    long nd_plen;
+    int nd_plen;
     struct RNode *nd_next;
 } rb_node_args_aux_t;
